Support markdown and navigation
Added the '@uiw/react-md-editor' dependency to allow us to edit and
parse markdown files. While not necessary for editing, a parser is
required to render markdown content. This package conveniently has both
an interactive editor and the markdown renderer. At the very least it is
a strong first-step.

As part of the changes, nav support had to be added so that I could
switch to the editor / reader views. This current nav is incredibly
basic, using three unformatted buttons; however, it is functional.