Cleanup Renderer File Management, Support Save New
Created a singleton `FileManager` class to facilitate communication
between IPC listeners and the markdown editor / reader. While still not
a perfect solution, this does cleanup the App immensely, moves the
listener creation to be on the file-scope instead of the component, and
moves text control almost entirely within the component it pertains to.

Also added some simple logic so that a user can save (CTRL + S) on a new
document instead of requiring save-as (CTRL + SHIFT + S).

Additionally, created a simple mock for `global.window.electron` since
Jest does not initialize this.